PlayerComponent handles interactions with hostiles
